ASM-3855 Re-factoring iom_8x4 SAN switch puppet module using puppet apply
