+++ lib/puppet/iom_8x4_model.rb
@@ -1,5 +1,7 @@
-module Puppet::Iom_8x4_Model
-  MODEL_HASH = {
-    '75.x' => 'Dell IOM 8|4'
-   }
+module Puppet
+	class Iom_8x4_Model
+	  MODEL_HASH = {
+	    '75.x' => 'Dell IOM 8|4'
+	   }
+	end
 end
