--- lib/puppet/util/network_device/iom_8x4.rb
@@ -1,5 +1,0 @@
-require 'puppet/util/network_device'
-
-module Puppet::Util::NetworkDevice::Iom_8x4
-end
-
