--- lib/puppet/util/network_device/iom_8x4/possible_facts.rb
@@ -1,3 +1,0 @@
-module Puppet::Util::NetworkDevice::Iom_8x4::PossibleFacts
-end
-
