--- lib/puppet/util/network_device/transport_iom_8x4.rb
@@ -1,4 +1,0 @@
-require 'puppet/util/network_device'
-
-module Puppet::Util::NetworkDevice::Transport_Iom_8x4
-end
