+++ lib/puppet_x/iom_8x4/possible_facts.rb
@@ -1,0 +1,7 @@
+#This class provides generic method to parse the fact from command output.
+module PuppetX
+  module Iom8x4
+    module PossibleFacts
+    end
+  end
+end
